apps/server/autocomplete/requirements.txt
python-dotenv
openai
PyMuPDF==1.23.8
requests
supabase
tiktoken
fastapi
langchain
langchain-openai
langchain-community
pydantic
sentence-transformers
tavily-python
PyPDF2
python-multipart
uvicorn
gunicorn
urllib3==1.26.15
baml-py